+++ src/test/java/io/github/kxng0109/aegisgate/AegisGateApplicationTests.java
@@ -16,7 +16,7 @@ class AegisGateApplicationTests {
 	@Container
 	@ServiceConnection
 	static final PostgreSQLContainer POSTGRES =
-			new PostgreSQLContainer(DockerImageName.parse("postgres:16-alpine"));
+			new PostgreSQLContainer(DockerImageName.parse("postgres:16.15-alpine"));
 
 	@Container
 	@ServiceConnection