+++ src/test/java/io/github/kxng0109/aegisgate/HighThroughputProfileTest.java
@@ -32,7 +32,7 @@ class HighThroughputProfileTest {
 	@Container
 	@ServiceConnection
 	static final PostgreSQLContainer POSTGRES =
-			new PostgreSQLContainer(DockerImageName.parse("postgres:16-alpine"));
+			new PostgreSQLContainer(DockerImageName.parse("postgres:16.15-alpine"));
 
 	@Container
 	@ServiceConnection