+++ src/test/java/io/github/kxng0109/aegisgate/contracts/GatewayPropertiesBindingTest.java
@@ -26,7 +26,7 @@ class GatewayPropertiesBindingTest {
 	@Container
 	@ServiceConnection
 	static final PostgreSQLContainer POSTGRES =
-			new PostgreSQLContainer(DockerImageName.parse("postgres:16-alpine"));
+			new PostgreSQLContainer(DockerImageName.parse("postgres:16.15-alpine"));
 
 	@Container
 	@ServiceConnection