+++ src/test/java/io/github/kxng0109/aegisgate/ledger/LedgerStressAndBackpressureIntegrationTest.java
@@ -46,7 +46,7 @@ class LedgerStressAndBackpressureIntegrationTest {
 
 	@Container
 	@ServiceConnection
-	static final PostgreSQLContainer POSTGRES = new PostgreSQLContainer("postgres:16-alpine");
+	static final PostgreSQLContainer POSTGRES = new PostgreSQLContainer("postgres:16.15-alpine");
 
 	@Autowired
 	private ApplicationEventPublisher eventPublisher;