+++ src/test/java/io/github/kxng0109/aegisgate/proxy/sse/SseFlushAutoConfigTest.java
@@ -28,7 +28,7 @@ class SseFlushAutoConfigTest {
 	@Container
 	@ServiceConnection
 	static final PostgreSQLContainer POSTGRES =
-			new PostgreSQLContainer(DockerImageName.parse("postgres:16-alpine"));
+			new PostgreSQLContainer(DockerImageName.parse("postgres:16.15-alpine"));
 
 	@Container
 	@ServiceConnection